Refactor hotels activity
